fix ecall SYS_EXEC in S-Mode
When detect an invalid SBI call number, the legacy BBL will delegate this trap back to S-Mode, but the latest BBL and OpenSBI will just return an error.
In the past, kernel_execve uses 'ecall' to trigger syscall from S-Mode (derived from 'int' on i386), then BBL handles this trap and returns to S-Mode handler. Now we have to simulate a trap from S-Mode: set CSRs and directly 'j'ump to the trap handler.